+++ lib/capybara/webkit/browser.rb
@@ -102,28 +102,76 @@ module Capybara::Webkit
       command("WindowFocus", selector)
     end
 
+    def window_open
+      command("WindowOpen")
+    end
+
+    def window_close(selector)
+      command("WindowClose", selector)
+    end
+
+    def window_resize(handle, width, height)
+      command("WindowResize", handle, width.to_i, height.to_i)
+    end
+
+    def window_size(handle)
+      JSON.parse(command("WindowSize", handle))
+    end
+
+    def window_maximize(handle)
+      command("WindowMaximize", handle)
+    end
+
     def get_window_handles
       JSON.parse(command('GetWindowHandles'))
     end
 
-    alias_method :window_handles, :get_window_handles
+    def window_handles
+      warn '[DEPRECATION] Capybara::Webkit::Browser#window_handles ' \
+        'is deprecated. Please use Capybara::Session#windows instead.'
+      get_window_handles
+    end
 
     def get_window_handle
       command('GetWindowHandle')
     end
 
-    alias_method :window_handle, :get_window_handle
+    def window_handle
+      warn '[DEPRECATION] Capybara::Webkit::Browser#window_handle ' \
+        'is deprecated. Please use Capybara::Session#current_window instead.'
+      get_window_handle
+    end
+
+    def accept_confirm(options)
+      command("SetConfirmAction", "Yes", options[:text])
+    end
 
     def accept_js_confirms
       command("SetConfirmAction", "Yes")
+    end
+
+    def reject_confirm(options)
+      command("SetConfirmAction", "No", options[:text])
     end
 
     def reject_js_confirms
       command("SetConfirmAction", "No")
     end
 
+    def accept_prompt(options)
+      if options[:with]
+        command("SetPromptAction", "Yes", options[:text], options[:with])
+      else
+        command("SetPromptAction", "Yes", options[:text])
+      end
+    end
+
     def accept_js_prompts
       command("SetPromptAction", "Yes")
+    end
+
+    def reject_prompt(options)
+      command("SetPromptAction", "No", options[:text])
     end
 
     def reject_js_prompts
@@ -136,6 +184,14 @@ module Capybara::Webkit
 
     def clear_prompt_text
       command("ClearPromptText")
+    end
+
+    def accept_alert(options)
+      command("AcceptAlert", options[:text])
+    end
+
+    def find_modal(id)
+      command("FindModal", id)
     end
 
     def url_blacklist=(black_list)
@@ -195,12 +251,16 @@ module Capybara::Webkit
       command("SetProxy")
     end
 
-    def resize_window(width, height)
-      command("ResizeWindow", width.to_i, height.to_i)
-    end
-
     def version
       command("Version")
+    end
+
+    def go_back
+      command("GoBack")
+    end
+
+    def go_forward
+      command("GoForward")
     end
 
     private