+++ lib/capybara/webkit/driver.rb
@@ -69,14 +69,22 @@ module Capybara::Webkit
     end
 
     def alert_messages
+      warn '[DEPRECATION] Capybara::Webkit::Driver#alert_messages ' \
+        'is deprecated. Please use Capybara::Session#accept_alert instead.'
       browser.alert_messages
     end
 
     def confirm_messages
+      warn '[DEPRECATION] Capybara::Webkit::Driver#confirm_messages ' \
+        'is deprecated. Please use Capybara::Session#accept_confirm ' \
+        'or Capybara::Session#dismiss_confirm instead.'
       browser.confirm_messages
     end
 
     def prompt_messages
+      warn '[DEPRECATION] Capybara::Webkit::Driver#prompt_messages ' \
+        'is deprecated. Please use Capybara::Session#accept_prompt ' \
+        'or Capybara::Session#dismiss_prompt instead.'
       browser.prompt_messages
     end
 
@@ -89,7 +97,17 @@ module Capybara::Webkit
     end
 
     def resize_window(width, height)
-      browser.resize_window(width, height)
+      warn '[DEPRECATION] Capybara::Webkit::Driver#resize_window ' \
+        'is deprecated. Please use Capybara::Window#resize_to instead.'
+      resize_window_to(current_window_handle, width, height)
+    end
+
+    def resize_window_to(handle, width, height)
+      browser.window_resize(handle, width, height)
+    end
+
+    def window_size(handle)
+      browser.window_size(handle)
     end
 
     def within_frame(selector)
@@ -102,8 +120,8 @@ module Capybara::Webkit
     end
 
     def within_window(selector)
-      current_window = window_handle
-      browser.window_focus(selector)
+      current_window = current_window_handle
+      switch_to_window(selector)
       begin
         yield
       ensure
@@ -111,36 +129,102 @@ module Capybara::Webkit
       end
     end
 
+    def switch_to_window(selector)
+      browser.window_focus(selector)
+    end
+
     def window_handles
       browser.get_window_handles
     end
 
-    def window_handle
+    def current_window_handle
       browser.get_window_handle
     end
 
+    def open_new_window
+      browser.window_open
+    end
+
+    def close_window(selector)
+      browser.window_close(selector)
+    end
+
+    def maximize_window(selector)
+      browser.window_maximize(selector)
+    end
+
     def accept_js_confirms!
+      warn '[DEPRECATION] Capybara::Webkit::Driver#accept_js_confirms! ' \
+        'is deprecated. Please use Capybara::Session#accept_confirm instead.'
       browser.accept_js_confirms
     end
 
     def dismiss_js_confirms!
+      warn '[DEPRECATION] Capybara::Webkit::Driver#dismiss_js_confirms! ' \
+        'is deprecated. Please use Capybara::Session#dismiss_confirm instead.'
       browser.reject_js_confirms
     end
 
     def accept_js_prompts!
+      warn '[DEPRECATION] Capybara::Webkit::Driver#accept_js_prompts! ' \
+        'is deprecated. Please use Capybara::Session#accept_prompt instead.'
       browser.accept_js_prompts
     end
 
     def dismiss_js_prompts!
+      warn '[DEPRECATION] Capybara::Webkit::Driver#dismiss_js_prompts! ' \
+        'is deprecated. Please use Capybara::Session#dismiss_prompt instead.'
       browser.reject_js_prompts
     end
 
     def js_prompt_input=(value)
+      warn '[DEPRECATION] Capybara::Webkit::Driver#js_prompt_input= ' \
+        'is deprecated. Please use Capybara::Session#accept_prompt instead.'
       if value.nil?
         browser.clear_prompt_text
       else
         browser.set_prompt_text_to(value)
       end
+    end
+
+    def go_back
+      browser.go_back
+    end
+
+    def go_forward
+      browser.go_forward
+    end
+
+    def accept_modal(type, options={})
+      options = modal_action_options_for_browser(options)
+
+      case type
+      when :confirm
+        id = browser.accept_confirm(options)
+      when :prompt
+        id = browser.accept_prompt(options)
+      else
+        id = browser.accept_alert(options)
+      end
+
+      yield
+
+      find_modal(type, id, options)
+    end
+
+    def dismiss_modal(type, options={})
+      options = modal_action_options_for_browser(options)
+
+      case type
+      when :confirm
+        id = browser.reject_confirm(options)
+      else
+        id = browser.reject_prompt(options)
+      end
+
+      yield
+
+      find_modal(type, id, options)
     end
 
     def wait?
@@ -174,12 +258,38 @@ module Capybara::Webkit
       [Capybara::Webkit::ClickFailed]
     end
 
+    def no_such_window_error
+      Capybara::Webkit::NoSuchWindowError
+    end
+
     def version
       [
         "Capybara: #{Capybara::VERSION}",
         "capybara-webkit: #{Capybara::Driver::Webkit::VERSION}",
         browser.version
       ].join("\n")
+    end
+
+    private
+
+    def modal_action_options_for_browser(options)
+      if options[:text].is_a?(Regexp)
+        options.merge(text: options[:text].source)
+      else
+        options.merge(text: Regexp.escape(options[:text].to_s))
+      end.merge(original_text: options[:text])
+    end
+
+    def find_modal(type, id, options)
+      Timeout::timeout(options[:wait] || Capybara.default_wait_time) do
+        browser.find_modal(id)
+      end
+    rescue ModalNotFound
+      raise Capybara::ModalNotFound,
+        "Unable to find modal dialog#{" with #{options[:original_text]}" if options[:original_text]}"
+    rescue Timeout::Error
+      raise Capybara::ModalNotFound,
+        "Timed out waiting for modal dialog#{" with #{options[:original_text]}" if options[:original_text]}"
     end
   end
 end