+++ lib/capybara/webkit/errors.rb
@@ -14,6 +14,15 @@ module Capybara::Webkit
   class TimeoutError < Timeout::Error
   end
 
+  class NoSuchWindowError < StandardError
+  end
+
+  class ConnectionError < StandardError
+  end
+
+  class ModalNotFound < StandardError
+  end
+
   class JsonError
     def initialize(response)
       error = JSON.parse response