+++ spec/selenium_compatibility_spec.rb
@@ -3,13 +3,15 @@ require 'spec_helper'
 describe Capybara::Webkit, 'compatibility with selenium' do
   include AppRunner
 
-  it 'generates the same events as selenium when filling out forms' do
+  it 'generates the same events as selenium when filling out forms', selenium_compatibility: true do
     run_application_for_html(<<-HTML)
       <html><body>
         <form onsubmit="return false">
           <label for="one">One</label><input type="text" name="one" id="one" />
           <label for="two">Two</label><input type="text" name="two" id="two" />
           <label for="three">Three</label><input type="text" name="three" id="three" readonly="readonly" />
+          <label for="textarea">Textarea</label>
+          <textarea name="textarea" id="textarea"></textarea>
           <input type="submit" value="Submit" id="submit" />
         </form>
         <script type="text/javascript">
@@ -35,6 +37,7 @@ describe Capybara::Webkit, 'compatibility with selenium' do
       fill_in "One", :with => "a new value"
       fill_in "Two", :with => "other value"
       fill_in "Three", :with => "readonly value"
+      fill_in "Textarea", :with => "last value"
       click_button "Submit"
     end
   end