+++ src/WebPage.h
@@ -10,6 +10,7 @@
 class WebPageManager;
 class InvocationResult;
 class NetworkReplyProxy;
+class QWebView;
 
 class WebPage : public QWebPage {
   Q_OBJECT
@@ -22,8 +23,12 @@ class WebPage : public QWebPage {
     QString userAgentForUrl(const QUrl &url ) const;
     void setUserAgent(QString userAgent);
     void setConfirmAction(QString action);
+    QString setConfirmAction(QString action, QString message);
+    QString setPromptAction(QString action, QString message, QString response);
+    QString setPromptAction(QString action, QString message);
     void setPromptAction(QString action);
     void setPromptText(QString action);
+    QString acceptAlert(QString);
     int getLastStatus();
     void setCustomNetworkAccessManager();
     bool render(const QString &fileName, const QSize &minimumSize);
@@ -33,7 +38,7 @@ class WebPage : public QWebPage {
     QVariantList alertMessages();
     QVariantList confirmMessages();
     QVariantList promptMessages();
-    void resetWindowSize();
+    void createWindow();
     void resetLocalStorage();
     QWebPage *createWindow(WebWindowType type);
     QString uuid();
@@ -46,6 +51,9 @@ class WebPage : public QWebPage {
     QString contentType();
     void mouseEvent(QEvent::Type type, const QPoint &position, Qt::MouseButton button);
     bool clickTest(QWebElement element, int absoluteX, int absoluteY);
+    void resize(int, int);
+    int modalCount();
+    QString modalMessage(int);
 
   public slots:
     bool shouldInterruptJavaScript();
@@ -57,11 +65,13 @@ class WebPage : public QWebPage {
     void handleSslErrorsForReply(QNetworkReply *reply, const QList<QSslError> &);
     void handleUnsupportedContent(QNetworkReply *reply);
     void replyFinished(QUrl &, QNetworkReply *);
+    void remove();
 
   signals:
     void pageFinished(bool);
     void requestCreated(QByteArray &url, QNetworkReply *reply);
     void replyFinished(QNetworkReply *reply);
+    void modalReady(int);
 
   protected:
     virtual void javaScriptConsoleMessage(const QString &message, int lineNumber, const QString &sourceID);
@@ -79,8 +89,8 @@ class WebPage : public QWebPage {
     QStringList getAttachedFileNames();
     void loadJavascript();
     void setUserStylesheet();
-    bool m_confirm;
-    bool m_prompt;
+    bool m_confirmAction;
+    bool m_promptAction;
     QVariantList m_consoleMessages;
     QVariantList m_alertMessages;
     QVariantList m_confirmMessages;
@@ -91,6 +101,9 @@ class WebPage : public QWebPage {
     QString m_errorPageMessage;
     void setFrameProperties(QWebFrame *, QUrl &, NetworkReplyProxy *);
     QPoint m_mousePosition;
+    QList<QVariantMap> m_modalResponses;
+    QStringList m_modalMessages;
+    void addModalMessage(bool, const QString &, const QRegExp &);
 };
 
 #endif //_WEBPAGE_H