+++ src/capybara.js
@@ -66,7 +66,9 @@ Capybara = {
   text: function (index) {
     var node = this.getNode(index);
     var type = (node.type || node.tagName).toLowerCase();
-    if (type == "textarea") {
+    if (!this.isNodeVisible(node)) {
+      return '';
+    } else if (type == "textarea") {
       return node.innerHTML;
     } else {
       return node.innerText || node.textContent;
@@ -270,8 +272,6 @@ Capybara = {
     textTypes = ["email", "number", "password", "search", "tel", "text", "textarea", "url"];
 
     if (textTypes.indexOf(type) != -1) {
-      this.focus(index);
-
       maxLength = this.attribute(index, "maxlength");
       if (maxLength && value.length > maxLength) {
         length = maxLength;
@@ -279,24 +279,41 @@ Capybara = {
         length = value.length;
       }
 
-      if (!node.readOnly)
+      if (!node.readOnly) {
+        this.focus(index);
+
         node.value = "";
 
-      for (strindex = 0; strindex < length; strindex++) {
-        CapybaraInvocation.keypress(value[strindex]);
-      }
+        for (strindex = 0; strindex < length; strindex++) {
+          CapybaraInvocation.keypress(value[strindex]);
+        }
 
+        if (value == '')
+          this.trigger(index, "change");
+      }
     } else if (type === "checkbox" || type === "radio") {
       if (node.checked != (value === "true")) {
         this.leftClick(index);
       }
-
     } else if (type === "file") {
       this.attachedFiles = Array.prototype.slice.call(arguments, 1);
       this.leftClick(index);
-
+    } else if (this.isContentEditable(node)) {
+      var content = document.createTextNode(value);
+      node.innerHTML = '';
+      node.appendChild(content);
     } else {
       node.value = value;
+    }
+  },
+
+  isContentEditable: function(node) {
+    if (node.contentEditable == 'true') {
+      return true;
+    } else if (node.contentEditable == 'false') {
+      return false;
+    } else if (node.contentEditable == 'inherit') {
+      return this.isContentEditable(node.parentNode);
     }
   },