+++ NEWS.md
@@ -1,3 +1,14 @@
+New for 1.3.0:
+
+* Capybara 2.4 compatibility.
+* Raise better errors if server fails to start
+* Offline application cache support.
+* Wildcard URL blacklist support.
+
+New for 1.2.0:
+* Capybara 2.3 compatibility.
+* Kill webkit_server when parent process closes stdin.
+
 New for 1.1.1:
 * Lock capybara dependency to < 2.2.0.